requirements.txt
PyYAML==6.0.3
requests==2.28.1
